requirements.txt
APScheduler==3.6.1
Click==7.0
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-SocketIO==4.2.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.2
pyzmq==18.1.0
six==1.12.0
tzlocal==2.0.0
Werkzeug==0.16.0
